Add flag to turn off non-blind cosmology messages
Picca will briefly become a dependency of vega (as I will discuss tomorrow), and we will be calling the Cosmo class from outside. Therefore we need a flag to turn-off the print statements done when there is no blinding, as they are not relevant when this is called from vega.